add brand in the project
- add test class for all html file
see #21
